requirements.txt

amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.1
celery==5.3.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comtypes==1.4.10
cryptography==44.0.2
distro==1.9.0
dj-database-url==2.3.0
Django==4.2.20
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==23.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.2.2
filelock==3.18.0
fsspec==2025.3.0
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
kombu==5.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.2
openai==1.68.2
packaging==24.2
pandas==2.1.4
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMuPDF==1.25.4
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyttsx3==2.98
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.1
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.2
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.1
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
